refactor: change pagination into ui from model
