--- src/components/model/Pagination/presentations/error.tsx
@@ -1,7 +1,0 @@
-import type { FC } from 'react';
-
-export const PaginationErrorPresentation: FC = () => (
-  <div className="w-full h-12 items-center justify-center flex">
-    ページ切り替えボタンの表示に失敗しました
-  </div>
-);
